Adjust gyoji dir to use cogs for command grouping.
Use of cogs allows for modular loading of functionality, at present we
are blindly loading all commands under the `gyoji/commands/` directory
but this is fine for now.

Slohyo related functionality is now in the `Slohyo` cog, and logging has
been moved to the `__init__` module within subdirectories. Removed the
bad resolution of users using a `UserResolver` and use a placeholder for
now.